// widgets/password-input/typings/PasswordInputProps.d.ts
/**
 * This file was generated from PasswordInput.xml
 * WARNING: All changes made to this file will be overwritten
 * @author Mendix Widgets Framework Team
 */
import { CSSProperties } from "react";
import { ActionValue, EditableValue } from "mendix";

export type SizeEnum = "xs" | "sm" | "md" | "lg" | "xl";

export type RadiusEnum = "xs" | "sm" | "md" | "lg" | "xl";

export interface PasswordInputContainerProps {
    name: string;
    class: string;
    style?: CSSProperties;
    tabIndex?: number;
    value: EditableValue<string>;
    label: string;
    description: string;
    placeholder: string;
    required: boolean;
    disabled: boolean;
    size: SizeEnum;
    radius: RadiusEnum;
    onChange?: ActionValue;
}

export interface PasswordInputPreviewProps {
    /**
     * @deprecated Deprecated since version 9.18.0. Please use class property instead.
     */
    className: string;
    class: string;
    style: string;
    styleObject?: CSSProperties;
    readOnly: boolean;
    renderMode: "design" | "xray" | "structure";
    translate: (text: string) => string;
    value: string;
    label: string;
    description: string;
    placeholder: string;
    required: boolean;
    disabled: boolean;
    size: SizeEnum;
    radius: RadiusEnum;
    onChange: {} | null;
}